Move the APIResource interface into its own file
